ft: PoW as a password-based key derivation fun. (see #1) [+]
proof-of-work is used for the hardening of the encryption process.
AES-256 bits (32 bytes) is quantum-safe (by now?), but the password
as defined by the user/client might be vulnerable to brute-force and
dictionary attacks. therefore, the proof-of-work ensures that the
derived encryption key from the password plain text takes a long time,
thus, slowing down such attacks. it turns the client encryption process
really slow, a good approach might be to store such encrypted private
keys with a less CPU-expensive difficulty, but whenever the user wants
to backup or restore her private keys outside her computer, she must
pass through this more CPU-expensive difficulty.

a planned / future command line interface may deal with that, for instance,
to load encrypted private keys from filesystem, the command-line may take
30-60 seconds by some certain PoW's difficulty, but whenever the user
wants to dump or load base-64 versions of the encrypted private keys,
5-6 minutes is a good/secure backup time (really? must re-check that --)

Signed-off-by: Marco Aurélio da Silva <marcoonroad@gmail.com>